refactor(phase1b): library.caching を train_util.py から分離
VAE latent / text encoder output のキャッシュ関連 8 関数を新規モジュール
library/caching.py に移動。train_util.py 側は後方互換のため re-export。

PR-1（dataset 系）の 2 つ目の PR。BaseDataset がこれらの caching 関数を
内部で呼び出すため、後続の dataset.py 抽出（PR-1c）に先立って caching を
独立させ、循環依存を回避する。

- library/caching.py: 新規 8 関数
  - is_disk_cached_latents_is_expected
  - trim_and_resize_if_required
  - load_images_and_masks_for_caching
  - cache_batch_latents
  - cache_batch_text_encoder_outputs
  - cache_batch_text_encoder_outputs_sd3
  - save_text_encoder_outputs_to_disk
  - load_text_encoder_outputs_from_disk
- library/train_util.py: 上記 8 関数の定義を削除し re-export に置換
- library/strategy_base.py: caching.load_images_and_masks_for_caching を直接 import
- finetune/prepare_buckets_latents.py: caching.* を直接 import

train_util.py には BucketManager / load_image / HIGH_VRAM / get_hidden_states_sdxl
が残るため、caching は循環回避のため関数内で `_tu()` 経由で参照する
（PR-1c で dataset.py へ移動した後に整理予定）。

動作仕様の変更なし。両方向の import 順序で動作確認済み。
